ci(actions): make tests run when changing dependencies
This change is necessary because previously changing the requirements would
not trigger the tests, but as shown by #7, this can catch us off-guard.

In that PR in particular, a downgrade in numpy, which did not trigger tests
has changed the API of some numpy functions used by `plaquette.pauli`, making
some tests fail due to the new (old) dependency.